+++ phpunit.xml
@@ -14,8 +14,7 @@
     </php>
   <testsuites>
     <testsuite name="OpenCage geocoder Test Suite">
-      <directory suffix=".php">./tests/</directory>
-      <exclude>tests/bootstrap.php</exclude>
+      <directory suffix="Test.php">./tests/</directory>
     </testsuite>
   </testsuites>
   <source>
